Suggested Changes as per PR review
