+++ app/src/main/java/com/example/ghclone/di/DataStoreModule.kt
@@ -7,11 +7,13 @@ import androidx.datastore.preferences.preferencesDataStoreFile
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
+private const val DATA_STORE_PREF_NAME = "settings"
+
 val DataStoreModule = module {
     single {
         PreferenceDataStoreFactory.create(
             corruptionHandler = ReplaceFileCorruptionHandler(produceNewData = { emptyPreferences() }),
-            produceFile = { androidContext().preferencesDataStoreFile("settings") },
+            produceFile = { androidContext().preferencesDataStoreFile(DATA_STORE_PREF_NAME) },
         )
     }
 }
