+++ app/src/main/java/com/example/ghclone/navigation/BottomNavItem.kt
@@ -3,29 +3,29 @@ package com.example.ghclone.navigation
 import com.example.ghclone.R
 
 data class BottomNavItem(
-    val label: String,
+    val label: Int,
     val icon: Int,
     val route: String,
 )
 
 val BottomNavItems = listOf(
     BottomNavItem(
-        label = "Home",
+        label = R.string.tab_home,
         icon = R.drawable.ic_tab_home,
         route = AppNav.HomeScreen.route,
     ),
     BottomNavItem(
-        label = "Notifications",
+        label = R.string.tab_notifications,
         icon = R.drawable.ic_tab_notification,
         route = AppNav.NotificationScreen.route,
     ),
     BottomNavItem(
-        label = "Explore",
+        label = R.string.tab_explore,
         icon = R.drawable.ic_tab_explore,
         route = AppNav.ExploreScreen.route,
     ),
     BottomNavItem(
-        label = "Profile",
+        label = R.string.tab_profile,
         icon = R.drawable.ic_tab_person,
         route = AppNav.ProfileScreen.route,
     ),
