+++ data/src/main/java/com/example/data/Constants.kt
@@ -10,4 +10,9 @@ object Constants {
         "https://github.com/login/oauth/authorize?client_id=${BuildConfig.ghClientID}" +
             "&scope=repo%20user:email%20read:user&redirect_uri=${BuildConfig.ghCallbackUrl}"
     const val AccessTokenUrl = "https://github.com/login/oauth/access_token"
+
+    const val CLIENT_ID_PARAM = "client_id"
+    const val CLIENT_SECRET_PARAM = "client_secret"
+    const val CODE_PARAM = "code"
+    const val REDIRECT_URI_PARAM = "redirect_uri"
 }
